Assemble shared utility functions
