+++ pkg/volume/csi/BUILD
@@ -36,6 +36,7 @@ go_test(
         "csi_client_test.go",
         "csi_mounter_test.go",
         "csi_plugin_test.go",
+        "csi_util_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
