+++ pkg/volume/csi/csi_mounter.go
@@ -18,9 +18,7 @@ package csi
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"os"
 	"path"
 
@@ -30,29 +28,11 @@ import (
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
-	kstrings "k8s.io/kubernetes/pkg/util/strings"
 	"k8s.io/kubernetes/pkg/volume"
 	"k8s.io/kubernetes/pkg/volume/util"
 )
 
 const defaultFSType = "ext4"
-
-//TODO (vladimirvivien) move this in a central loc later
-var (
-	volDataKey = struct {
-		specVolID,
-		volHandle,
-		driverName,
-		nodeName,
-		attachmentID string
-	}{
-		"specVolID",
-		"volumeHandle",
-		"driverName",
-		"nodeName",
-		"attachmentID",
-	}
-)
 
 type csiMountMgr struct {
 	k8s          kubernetes.Interface
@@ -77,11 +57,6 @@ func (c *csiMountMgr) GetPath() string {
 	dir := path.Join(getTargetPath(c.podUID, c.specVolumeID, c.plugin.host), "/mount")
 	glog.V(4).Info(log("mounter.GetPath generated [%s]", dir))
 	return dir
-}
-
-func getTargetPath(uid types.UID, specVolumeID string, host volume.VolumeHost) string {
-	specVolID := kstrings.EscapeQualifiedNameForDisk(specVolumeID)
-	return host.GetPodVolumeDir(uid, kstrings.EscapeQualifiedNameForDisk(csiPluginName), specVolID)
 }
 
 // volume.Mounter methods
@@ -267,10 +242,7 @@ func (c *csiMountMgr) TearDownAt(dir string) error {
 	driverName := data[volDataKey.driverName]
 
 	if c.csiClient == nil {
-		addr := fmt.Sprintf(csiAddrTemplate, driverName)
-		client := newCsiDriverClient("unix", addr)
-		glog.V(4).Infof(log("unmounter csiClient setup [volume=%v,driver=%v]", volID, driverName))
-		c.csiClient = client
+		c.csiClient = getCSIClient(driverName)
 	}
 
 	ctx, cancel := context.WithTimeout(context.Background(), csiTimeout)
@@ -291,50 +263,6 @@ func (c *csiMountMgr) TearDownAt(dir string) error {
 	glog.V(4).Infof(log("mounte.TearDownAt successfully unmounted dir [%s]", dir))
 
 	return nil
-}
-
-// saveVolumeData persists parameter data as json file using the location
-// generated by /var/lib/kubelet/pods/<podID>/volumes/kubernetes.io~csi/<specVolId>/volume_data.json
-func saveVolumeData(p *csiPlugin, podUID types.UID, specVolID string, data map[string]string) error {
-	dir := getTargetPath(podUID, specVolID, p.host)
-	dataFilePath := path.Join(dir, volDataFileName)
-
-	file, err := os.Create(dataFilePath)
-	if err != nil {
-		glog.Error(log("failed to save volume data file %s: %v", dataFilePath, err))
-		return err
-	}
-	defer file.Close()
-	if err := json.NewEncoder(file).Encode(data); err != nil {
-		glog.Error(log("failed to save volume data file %s: %v", dataFilePath, err))
-		return err
-	}
-	glog.V(4).Info(log("volume data file saved successfully [%s]", dataFilePath))
-	return nil
-}
-
-// loadVolumeData uses the directory returned by mounter.GetPath with value
-// /var/lib/kubelet/pods/<podID>/volumes/kubernetes.io~csi/<specVolumeId>/mount.
-// The function extracts specVolumeID and uses it to load the json data file from dir
-// /var/lib/kubelet/pods/<podID>/volumes/kubernetes.io~csi/<specVolId>/volume_data.json
-func loadVolumeData(dir string, fileName string) (map[string]string, error) {
-	// remove /mount at the end
-	dataFileName := path.Join(dir, fileName)
-	glog.V(4).Info(log("loading volume data file [%s]", dataFileName))
-
-	file, err := os.Open(dataFileName)
-	if err != nil {
-		glog.Error(log("failed to open volume data file [%s]: %v", dataFileName, err))
-		return nil, err
-	}
-	defer file.Close()
-	data := map[string]string{}
-	if err := json.NewDecoder(file).Decode(&data); err != nil {
-		glog.Error(log("failed to parse volume data file [%s]: %v", dataFileName, err))
-		return nil, err
-	}
-
-	return data, nil
 }
 
 // isDirMounted returns the !notMounted result from IsLikelyNotMountPoint check
